implemented the ability to edit and delete recipes

+++ 01-recipe_app/src/utils/LocalDB.ts
@@ -9,17 +9,40 @@ export default class LocalDB {
         for (let i = 0; i < window.localStorage.length; i++) {
             let key = window.localStorage.key(i)!;
 
-            if (key.startsWith('recipe-')) {
+            if (key.startsWith(this.recipePrefix)) {
                 recipes.push(JSON.parse(localStorage.getItem(key)!));
             }
         }
         return recipes;
     }
 
+    public getRecipe(id: string): Recipe | undefined {
+        let recipe = localStorage.getItem(this.recipePrefix + id);
+        if (recipe) {
+            return JSON.parse(recipe);
+        }
+        return undefined;
+    }
+
     public addRecipe(recipe: Recipe): void {
+        if (this.getRecipe(recipe.id)) { throw new Error('Recipe already exists')}
         window.localStorage.setItem(
-            this.recipePrefix + recipe.name,
+            this.recipePrefix + recipe.id,
             JSON.stringify(recipe)
         );
     }
+
+    public updateRecipe(recipe: Recipe): void {
+        if (!this.getRecipe(recipe.id)) { throw new Error('Recipe does not already exist')}
+        window.localStorage.setItem(
+            this.recipePrefix + recipe.id,
+            JSON.stringify(recipe)
+        );
+    }
+
+    public deleteRecipe(id: string): void {
+        localStorage.removeItem(this.recipePrefix + id);
+    }
+
+    
 }
